Перенёс все классы из папки Classes в три другие папки. итого получилось 4 папки: Models, View, ViewModels, Data. Начал перенос методов из одних классов в другие, лучше создам новую ветку, и уже в ней буду это делать.
