--- ConsoleApp2/Classes/SerializeDeserialize.cs
@@ -1,41 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.Json;
-using System.Threading.Tasks;
-
-namespace ConsoleApp2.Classes
-{
-    internal class SerializeDeserialize
-    {
-
-        public static bool TryDeserializ<T>(string path, ref T element)//вынести текст ошибок в свой класс ошибок
-            where T : new()
-        {
-            try
-            {
-                string textJson = File.ReadAllText(path);
-                element = JsonSerializer.Deserialize<T>(textJson)!;
-                var a = new T();
-                return true;
-            }
-            catch (FileNotFoundException)
-            {
-                Console.WriteLine("Не найден файл:\n{0}", path);
-                return false;
-            }
-            catch (JsonException)
-            {
-                Console.WriteLine("Ошибка в файле:\n{0}", path);
-                return false;
-            }
-            catch
-            {
-                Console.WriteLine("Ошибка, попробуйте еще раз.", path);
-                return false;
-            }
-        }
-
-    }
-}
