+++ ConsoleApp2/Models/Basket.cs
@@ -2,10 +2,12 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.Encodings.Web;
 using System.Text.Json;
 using System.Text.RegularExpressions;
+using System.Text.Unicode;
 using System.Threading.Tasks;
-using ConsoleApp2.Classes;
+using ConsoleApp2.Views;
 using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace ConsoleApp2.Models
@@ -21,43 +23,15 @@ namespace ConsoleApp2.Models
             Tickets.Add(new Ticket(filmScreening.Name, filmScreening.Date, filmScreening.Time, filmScreening.Price));
             Price += filmScreening.Price;
             NumberTickets++;
-            MessageTicketPurchased();
-        }
-        public static void MessageTicketPurchased()
-        {
-            Console.WriteLine("Билет куплен");
-        }
-        public static Basket BuyTickets(Dictionary<string, List<FilmScreening>> filmScreenings, FilmsPoster filmsPoster)
-        {
-            Basket basket = new();
-            bool flagBuyTickets = true;
-            while (flagBuyTickets)
-            {
-                // Выбор фильма
-                List<FilmScreening> filmScreeningsInOneFilm = FilmScreenings.ChooseFilmScreeingInCertainFilm(filmScreenings, filmsPoster);
-                // Выбор даты
-                List<FilmScreening> filmScreeningsInCertainDate = FilmScreenings.ChooseFilmScreeingInCertainDate(filmScreeningsInOneFilm);
-                // Выбор времени
-                FilmScreening filmScreeningInCertainTime = FilmScreening.ChooseFilmScreeningsInCertainTime(filmScreeningsInCertainDate);
-
-                if (FilmScreening.IsPlacesNotEmpty(filmScreeningInCertainTime))
-                {
-                    FilmScreening.OutputCountPlace(filmScreeningInCertainTime);
-                    if (FilmScreening.PoolYesOrNo("Купить билет"))
-                    {
-                        AddTicket(filmScreeningInCertainTime);
-                    }
-                    ConsoleMessages.MessageCheck();
-                    flagBuyTickets = !FilmScreening.PoolYesOrNo("Закончить");
-                }
-            }
-            return basket;
         }
         public static void Save(string path)
         {
-            var options = new JsonSerializerOptions { WriteIndented = true }; // опция для развертывания json файла
+            var options = new JsonSerializerOptions // опции для развертывания json файла и сериализации в Кириллицу
+            {
+                Encoder = JavaScriptEncoder.Create(UnicodeRanges.BasicLatin, UnicodeRanges.Cyrillic),
+                WriteIndented = true
+            };
             string jsonString = JsonSerializer.Serialize(Tickets, options); // список в строку
-            jsonString = Regex.Replace(jsonString, @"\\u([0-9A-Fa-f]{4})", m => "" + (char)Convert.ToInt32(m.Groups[1].Value, 16)); // меняем кодировку 
             File.WriteAllText(path, jsonString); // запись в файл .json
         }
     }
