+++ ConsoleApp2/Models/FilmsPoster.cs
@@ -2,9 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
-namespace ConsoleApp2.Classes
+namespace ConsoleApp2.Models
 {
     internal class FilmsPoster
     {
@@ -13,8 +12,8 @@ namespace ConsoleApp2.Classes
             Films = films;
             Count = films.Count;
         }
-        public List<Film> Films { get;}
-        public int Count { get; private set;}
+        public List<Film> Films { get; }
+        public int Count { get; private set; }
         public void MessageNamesAllFilms()
         {
             MessageFilmsAtTheBoxOffice();
